pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptguard"
version = "0.1.0a1"
description = "Local LLM proxy that prevents PII and sensitive data from leaving developer machines."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Mahdi Hedhli", email = "mhedhli@42-corp.com" }]
keywords = ["llm", "dlp", "pii", "proxy", "security", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "structlog>=24.1",
]

[project.optional-dependencies]
proxy = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]
opf-service = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "transformers>=4.44",
    "torch>=2.3",
]
presidio = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/MahdiHedhli/PromptGuard"
Issues = "https://github.com/MahdiHedhli/PromptGuard/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/promptguard"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
asyncio_mode = "auto"
# Hatchling's editable install writes _editable_impl_<pkg>.pth, which
# Python's site.py skips because the filename starts with an underscore.
# Adding `src` to pythonpath here makes pytest portable regardless of
# editable-install state (uv sync vs uv pip install -e .).
pythonpath = ["src"]
markers = [
    "integration: end-to-end tests that exercise the detection pipeline",
    "slow: tests that take more than ~1s",
    "docker: tests that require docker compose stack to be running",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
exclude = ["tests/", "benchmarks/"]
